Upgrade snekql 0.6 and migrate Python typing to ty (#688)
Closes #687.

## Summary

- upgrade the host to `snekql>=0.6.0` and migrate owner-qualified model
annotations and query types to the 0.6 API
- replace Pyright with strict `ty==0.0.75` validation in the host,
snekok, Docker image, task recipes, and contributor docs
- compose one ordered host migration declaration, preserving three
observed legacy history prefix sets and retired Proposal identities for
safe 0.6 adoption
- canonicalize historical Conversation schema defaults transactionally
after migration and cover fresh and legacy upgrade paths
- update test fixtures that previously composed independent migration
histories or used migrations for fixture data

## Validation

- `uv run ty check`
- `uv run ruff check .`
- `uv run ruff format --check .`
- `uv run python -m snektest tests/` (1,513 passed)
- snekok `ty`, Ruff, format, and tests (39 passed)
- agent typecheck, lint, format, and tests (34 passed)
- web typecheck, lint, format, and tests (408 passed)
- `just codegen-check`
- `just validate-web-smoke` (50 passed, 10 expected skips)
- upgraded copied databases carrying 107, 161, and 168 legacy migration
names to the 184-entry chain; Conversation row counts were preserved and
`Database.verify([Conversation])` passed

+++ apps/host/tether/background_runtime.py
@@ -440,9 +440,11 @@ class BackgroundRuntime:
                     )
                     raise
                 next_failure_count = consecutive_failures + 1
-                retry_seconds = min(
-                    self._retry_base_seconds * (2 ** (next_failure_count - 1)),
-                    self._retry_cap_seconds,
+                retry_seconds = float(
+                    min(
+                        self._retry_base_seconds * (2 ** (next_failure_count - 1)),
+                        self._retry_cap_seconds,
+                    )
                 )
                 self._health[registration.name] = replace(
                     self._health[registration.name],
